+++ components/info/impl/src/main/java/com/flipperdevices/info/impl/compose/navigation/NavGraphRoute.kt
@@ -1,0 +1,6 @@
+package com.flipperdevices.info.impl.compose.navigation
+
+enum class NavGraphRoute {
+    Info,
+    FullInfo
+}